+++ qcluster/features.py
@@ -1,27 +1,21 @@
 from sentence_transformers import SentenceTransformer
 import torch
-from datamodels import Samples, Instructions
+from datamodels import SampleCollection
 from qcluster import ROOT_DIR
 
-
-def create_instruction_embeddings(samples: Samples) -> torch.Tensor:
+def create_embeddings(texts: list[str],
+                      model: SentenceTransformer) -> torch.Tensor:
     """
-    Creates embeddings for the categories of the given samples.
+    Creates embeddings for the given texts using the specified model.
 
     Args:
-        samples: A Samples object containing the list of samples.
-
+        texts: A list of strings to be encoded.
+        model: A SentenceTransformer model used to encode the texts.
     Returns:
-        A torch.Tensor containing the embeddings for the categories.
+        A torch.Tensor containing the embeddings for the texts.
     """
-    model = SentenceTransformer('all-MiniLM-L6-v2')
-    instructions = (
-        Instructions
-        .from_samples(samples)
-        .to_list_of_strings()
-    )
-    assert isinstance(instructions[0], str)
-    embeddings = model.encode(instructions, convert_to_tensor=True)
+    assert isinstance(texts[0], str)
+    embeddings = model.encode(texts, convert_to_tensor=True)
     return embeddings
 
 
@@ -31,11 +25,10 @@ if __name__ == '__main__':
         / "data"
         / "Bitext_Sample_Customer_Support_Training_Dataset_27K_responses-v11.csv"
     )
-    samples = Samples.from_csv(csv_file_path)[:10]
-    # Generate embeddings
-    instruction_embeddings = create_instruction_embeddings(samples)
-
-    # Print the embeddings
+    MODEL = SentenceTransformer('all-MiniLM-L6-v2')
+    samples_ = SampleCollection.from_csv(csv_file_path)[:3]
+    instruction_embeddings = create_embeddings(
+        [i.instruction for i in samples_], MODEL)
     print("Instruction Embeddings:")
     print(instruction_embeddings)
     print(f"Shape of embeddings: {instruction_embeddings.shape}")